requirements.txt
fastapi
huggingface_hub
python-dotenv